Interpret time_since_last_lan_connection as int64
(cherry picked from commit 1e1d845fbc1afd093fa9994f7d0fd0b303185525)

+++ Connector.cpp
@@ -491,7 +491,7 @@ SStatusDetails Connector::getStatusDetails(SConnection* connection)
 			servers[i].name = json_servers[i]["name"].asString();
 		}
 		ret.servers=servers;
-		ret.time_since_last_lan_connection = root["time_since_last_lan_connection"].asInt();
+		ret.time_since_last_lan_connection = root["time_since_last_lan_connection"].asInt64();
 		ret.internet_connected = root["internet_connected"].asBool();
 		ret.internet_status = root["internet_status"].asString();
